Revert "Always cleanup GroupGuid/OriginalGroupGuid while set new group. possible fix bug-using problems, identified after 3023. testing."
This reverts commit 5fa8c94c83209f7288e64948ad7dfddda151ff2f.

+++ src/game/Player.cpp
@@ -20620,8 +20620,12 @@ void Player::SendPetComboPoints(Unit* pet, ObjectGuid targetGuid, uint8 combopoi
 
 void Player::SetGroup(ObjectGuid const& guid, int8 subgroup)
 {
-    m_groupGuid.Clear();
-    m_group.unlink();
+    if (guid.IsEmpty())
+    {
+        m_groupGuid.Clear();
+        m_group.unlink();
+        return;
+    }
 
     Group* group = sObjectMgr.GetGroup(guid);
 
@@ -20632,6 +20636,7 @@ void Player::SetGroup(ObjectGuid const& guid, int8 subgroup)
 
     // never use SetGroup without a subgroup unless you specify NULL for group
     MANGOS_ASSERT(subgroup >= 0);
+    m_group.unlink();
     m_group.link(group, this);
     m_group.setSubGroup((uint8)subgroup);
 }
@@ -21925,8 +21930,12 @@ void Player::RemoveFromBattleGroundRaid()
 
 void Player::SetOriginalGroup(ObjectGuid const& guid, int8 subgroup)
 {
-    m_originalGroupGuid.Clear();
-    m_originalGroup.unlink();
+    if (guid.IsEmpty())
+    {
+        m_originalGroupGuid.Clear();
+        m_originalGroup.unlink();
+        return;
+    }
 
     Group* group = sObjectMgr.GetGroup(guid);
 
@@ -21937,6 +21946,7 @@ void Player::SetOriginalGroup(ObjectGuid const& guid, int8 subgroup)
 
     // never use SetOriginalGroup without a subgroup unless you specify NULL for group
     MANGOS_ASSERT(subgroup >= 0);
+    m_originalGroup.unlink();
     m_originalGroup.link(group, this);
     m_originalGroup.setSubGroup((uint8)subgroup);
 }
